addonHandler: Reorganize code to improve modularity
Move the Addon class from addonHandler/__init__.py into its own module (addon.py) to improve code organization and readability. Also move AddonStateCategory enum and state management from addonStore/models/status.py to addonHandler/addonState.py where it more logically belongs.

This is a refactoring change that:
- Reduces the size of addonHandler/__init__.py by extracting Addon class
- Places addon state management closer to the addon code
- Improves overall code modularity and organization
