+++ source/addonStore/models/version.py
@@ -9,6 +9,7 @@ from typing import (
 	Optional,
 	Protocol,
 )
+
 import addonAPIVersion
 
 
@@ -51,18 +52,18 @@ class SupportsVersionCheck(Protocol):
 	@property
 	def _hasOverriddenCompat(self) -> bool:
 		"""If True, this add-on has been manually overriden. The affects of override may be pending restart"""
-		import addonHandler
-		from addonStore.models.status import AddonStateCategory
+		import addonHandler.addonState
+		from addonHandler.addonState import AddonStateCategory
 
 		return (
-			self.name in addonHandler.state[AddonStateCategory.OVERRIDE_COMPATIBILITY]
-			or self.name in addonHandler.state[AddonStateCategory.PENDING_OVERRIDE_COMPATIBILITY]
+			self.name in addonHandler.addonState.state[AddonStateCategory.OVERRIDE_COMPATIBILITY]
+			or self.name in addonHandler.addonState.state[AddonStateCategory.PENDING_OVERRIDE_COMPATIBILITY]
 		)
 
 	@property
 	def overrideIncompatibility(self) -> bool:
 		"""If True, NVDA should enable this add-on where it would normally be blocked due to incompatibility."""
-		from addonHandler import AddonStateCategory, state
+		from addonHandler.addonState import AddonStateCategory, state
 
 		return self.name in state[AddonStateCategory.OVERRIDE_COMPATIBILITY] and self.canOverrideCompatibility
 
@@ -71,7 +72,7 @@ class SupportsVersionCheck(Protocol):
 		Should be reset when changing to a new breaking release,
 		and when this add-on is updated, disabled or removed.
 		"""
-		from addonHandler import AddonStateCategory, state
+		from addonHandler.addonState import AddonStateCategory, state
 
 		if self.name in state[AddonStateCategory.PENDING_OVERRIDE_COMPATIBILITY]:
 			raise RuntimeError(f"{self.name} is already pending override compatibility.")
